requirements.txt
python-telegram-bot>=20.0
Pillow>=10.0.0
gTTS>=2.3.2
SpeechRecognition>=3.10.0
aiofiles>=23.0.0
aiohttp>=3.9.0
pytz>=2023.3
python-dateutil>=2.8.2
cryptography>=41.0.0
jsonschema>=4.19.0
pydub>=0.25.1
psutil>=5.9.0
colorlog>=6.7.0